Update Node Problems from previous merge with master.
Uses a screen width and size variable to stroe nodes in a 20 pixel area.

Fixed broken movement. Movement now occurs under placeholder destination coords set.

***TO COMPLETE***

Parts of map (Right side) are unaccessible to A*. This is due to a sizing issue.

Create a frequent nodes used for further searches of map.

Add a refill bar to the toilet once they have gone.