Added a dancefloor.
In the current state, the person moves to the dancefloor, gets assigned a random coord within the floor.

To Do:

GUI - Make dancefloor transparent

While on dancefloor - Delay & Dance and after a period of time move back to the idle state, refilling the dance meter.

Added a clear of the astar array when in dance state, so that when using astar again, they will not need to details of the original contents.

Agents still clip on some occasions, think this may be a flocking issue as per the print statements (Something about objects being within rejection?)